PR feedback + CI
